% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelsa_get.R
\name{download_chelsa}
\alias{download_chelsa}
\title{Download CHELSA V_2.1 data}
\usage{
download_chelsa(target_path, years, vars, months)
}
\arguments{
\item{target_path}{Download folder on local machine}

\item{years}{Vector containing the years from which monthly data should be downloaded (must be between 1972-2013)}

\item{vars}{Character string containing the CHELSA variables to be downloaded, can be one or more of "pr", "tas", "tasmax", "tasmin".}

\item{months}{Vector containing the months to be downloaded (1-12).}
}
\description{
This function downloads CHELSA V_2.1 climate layers (Karger et al., 2017) into a specified folder.
}
\examples{
vars <- c("pr", "tas", "tasmax", "tasmin")
months <- 1:12
years <- c(1979:1981)
path <- "mypath/"
download_chelsa(target_path, years, vars, months)

}
\references{
Karger, D., Conrad, O., Böhner, J. et al. Climatologies at high resolution for the earth’s land surface areas. Sci Data 4, 170122 (2017). https://doi.org/10.1038/sdata.2017.122
}
\author{
Simon Kapitza \email{simon.statecology@gmail.com}
}
