% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_layers.R
\name{fix_layers}
\alias{fix_layers}
\title{Load future predictions of bioclim variables.}
\usage{
fix_layers(mask_name, path)
}
\arguments{
\item{bnd}{\code{boundary} object.}

\item{var}{Variable. Valid names are 'tmin', 'tmax', 'prec' and 'bio'}

\item{res}{Resolution. Valid resolutions are 0.5, 2.5, 5 and 10.}

\item{rcp}{Representative Concenctration Pathway. Valid values are 26, 45, 60 and 85.}

\item{model}{GCM model under which predictions were made. Should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", "HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or "NO".}

\item{year}{Time step of predictions. Should be 50 or 70.}

\item{...}{Further parameters passed to raster::getData. See /code{?raster::getData}.}
}
\value{
raster object containing single layer or stack of GCM predictions of bioclim variables, cropped to extent of \code{bnd}.
}
\description{
This function downloads GCM of bioclim variables and crops them to the extent provided by boundary file.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
