% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tilenames.R
\name{get_tilenames}
\alias{get_tilenames}
\title{Get the names and center coordinates of worldclim tiles at 0.5 arc sec.}
\usage{
get_tilenames(bnd)
}
\arguments{
\item{bnd}{Spatial object from which extent can be extracted.}
}
\value{
List of tilenames in accordance with convention and their centre coordinates.
}
\description{
Function uses extent of input to identify which worldclim tiles are necessary to produce layers for study area.
}
\details{
The output can be used in \code{get_wctiles} to download set of worldclim tiles at 0.5 arc sec for merging to study area extent.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
