% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wctiles.R
\name{get_wctiles}
\alias{get_wctiles}
\title{Load worldclim data.}
\usage{
get_wctiles(tiles, var, path)
}
\arguments{
\item{tiles}{Character vector containing worldclim tile names (0.5 arcmin).}

\item{var}{Which worldclim variable? Valid names are 'tmin', 'tmax', 'prec' and 'bio'.}

\item{path}{target path where data is downloaded.}
}
\value{
raster object containing list of worldclim tiles.
}
\description{
This function is a wrapper around raster::getData. Provide a character vector with worldclim 0.5 arcmin tile names and those tiles will be downloaded.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
