% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_get.R
\name{tile_get}
\alias{tile_get}
\title{Load worldclim data.}
\usage{
tile_get(tiles, var, path)
}
\arguments{
\item{tiles}{Character vector containing worldclim tile names (0.5 arcmin).}

\item{var}{Which worldclim variable? Valid names are 'tmin', 'tmax', 'prec' and 'bio'.}

\item{path}{target path where data is downloaded.}
}
\value{
raster object containing list of worldclim tiles.
}
\description{
This function is a wrapper for \code{raster::getData}. Provide a character vector with WorldClim 0.5 arcmin tile names and those tiles will be downloaded.
}
\examples{
boundary <- getData("GADM", country = "FRA", level = 0) #Downloads France boundaries
tilenames <- tile_name(boundary) #Determines names of the worldclim tiles covering France
wctiles <- tile_get(tiles = tilenames, var = "bio") #downloads tiles covering France to working directory
}
\author{
Simon Kapitza \email{simon.statecology@gmail.com}
}
