% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_merge.R
\name{tile_merge}
\alias{tile_merge}
\title{Merge worldclim data.}
\usage{
tile_merge(biotiles)
}
\arguments{
\item{biotiles}{"List with 0.5 arc min WorldClim tile stacks (or single rasters) (i.e. produced by \code{tile_get}). Both stacks have to have the same variables in them.}
}
\value{
RasterStack or Raster containing merged WorldClim variable(s).
}
\description{
This function merges tiles downloaded from WorldClim into a single raster.
}
\details{
\code{raster::merge} is used for merging, at default settings. That means that in case of overlaps the cell values of the object that comes first in the sequence are retained. But this shouldn't matter much, because there are no overlaps between WorldClim tiles.
}
\examples{
boundary <- getData("GADM", country = "FRA", level = 0) #Downloads France boundaries
tilenames <- tile_name(boundary) #Determines names of the worldclim tiles covering France
wctiles <- tile_get(tiles = tilenames, var = "bio") #downloads tiles covering France to working directory
merged <- tile_merge(wctiles)
}
\author{
Simon Kapitza \email{simon.statecology@gmail.com}
}
