% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_name.R
\name{tile_name}
\alias{tile_name}
\title{Get the names of 0.5 arcmin WorldClim tiles.}
\usage{
tile_name(bnd)
}
\arguments{
\item{bnd}{Spatial feature or Raster object.}
}
\value{
Character vector of tilenames.
}
\description{
Provide projected spatial features of your study area. The function determines which WorldClim tiles they intersect with and returns the tile names as character vector.
}
\details{
Function uses input to identify which WorldClim tiles are necessary to produce layers for study area. The output can be used in \code{tile_get} to download set of worldclim tiles at 0.5 arc sec for merging to study area extent.
}
\examples{
boundary <- getData("GADM", country = "FRA", level = 0) #Downloads France boundaries
tilenames <- tile_name(boundary) #Determines names of the worldclim tiles covering France
}
\author{
Simon Kapitza \email{simon.statecology.gmail.com}
}
