% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_wctiles.R
\name{merge_wctiles}
\alias{merge_wctiles}
\title{Merge worldclim data.}
\usage{
merge_wctiles(biotiles)
}
\arguments{
\item{biotiles}{"List with 0.5 arc min worldclim tiles, either stacks or rasters. Variables may not vary across tiles.}
}
\value{
raster object containing single layer or stack of worldclim variables.
}
\description{
This function merges tiles downloaded from worldclim into a single raster.
}
\details{
List object required as input is prodcued by \code{get_wctile}.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
