% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_merge.R
\name{tile_merge}
\alias{tile_merge}
\title{Merge worldclim data.}
\usage{
tile_merge(tiles)
}
\arguments{
\item{tiles}{"List with 0.5 arc min WorldClim tile stacks (or single rasters) (i.e. produced by \code{tile_get}). Both stacks have to have the same variables in them.}
}
\value{
RasterStack or Raster containing merged WorldClim variable(s).
}
\description{
This function merges tiles downloaded from WorldClim into a single raster.
}
\details{
\code{raster::merge} is used for merging, at default settings. That means that in case of overlaps the cell values of the object that comes first in the sequence are retained. But this shouldn't matter much, because there are no overlaps between WorldClim tiles.
}
\examples{
boundary <- getData("GADM", country = "FRA", level = 0) #Downloads France boundaries

tilenames <- tile_name(boundary, name = 'worldclim') #Determines names of the worldclim tiles covering France
srtmtilenames <- tile_name(boundary, name = 'srtm') #Determines names of the srtm tiles covering France

wctiles <- tile_get(tiles = tilenames, name = 'srtm', var = "bio") #downloads WorldClim tiles covering France
srtmtiles <- tile_get(tiles = srtmtilenames, name = 'srtm') #downloads SRTM tiles covering France

wcmerged <- tile_merge(wctiles)
srtmmerged <- tile_merge(srtmtiles)
}
\author{
Simon Kapitza \email{simon.statecology@gmail.com}
}
