% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_merge.R
\name{tile_merge}
\alias{tile_merge}
\title{Merge worldclim data.}
\usage{
tile_merge(biotiles)
}
\arguments{
\item{biotiles}{"List with 0.5 arc min worldclim tile stacks (or single rasters) (i.e. produced by get_wctiles). Both stacks have to have the same variables in them.}
}
\value{
raster stack (or single raster) containing merged worldclim variable(s).
}
\description{
This function merges tiles downloaded from worldclim into a single raster.
}
\details{
\code{raster::merge} is used for merging, at default settings. That means that in case of overlaps the cell values of the object that comes first in the sequence are retained. But this shouldn't matter much, because there are no overlaps between worldclim tiles.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
