% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_name.R
\name{tile_name}
\alias{tile_name}
\title{Get the names and center coordinates of worldclim tiles at 0.5 arc sec.}
\usage{
tile_name(bnd)
}
\arguments{
\item{bnd}{Spatial object in latlong projection from which extent can be extracted.}
}
\value{
List with two elements. First element is Character vector of tilenames, second element contains tile centre coordinates.
}
\description{
Get the names and center coordinates of worldclim tiles at 0.5 arc sec.
}
\details{
Function uses extent of input to identify which worldclim tiles are necessary to produce layers for study area. The output can be used in \code{tile_get} to download set of worldclim tiles at 0.5 arc sec for merging to study area extent.
}
\author{
Simon Kapitza \email{kapitzas@student.unimelb.edu.au}
}
