% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_name.R
\name{tile_name}
\alias{tile_name}
\title{Get WorldClim and SRTM tile names.}
\usage{
tile_name(bnd, name = "worldclim")
}
\arguments{
\item{bnd}{Projected spatial feature or raster object.}

\item{name}{Valid values are 'worldclim' and 'srtm'. Default is 'worldclim'}
}
\value{
Character vector of tilenames of WorldClim or SRTM tiles.
}
\description{
Provide projected Spatial object (can be features or raster) and determine which WorldClim or SRTM tiles your study area intersects with.
}
\details{
Function uses the spatial object input to identify WorldClim or SRTM tiles to be downloaded. The output of this function can be used in \code{tile_get()} to download the according data via \code{raster::getData()}.
}
\examples{
boundary <- getData("GADM", country = "FRA", level = 0) #Downloads France boundaries

tilenames <- tile_name(boundary, name = 'worldclim') #Determines names of the worldclim tiles covering France
srtmtilenames <- tile_name(boundary, name = 'srtm') #Determines names of the srtm tiles covering France
}
\author{
Simon Kapitza \email{simon.statecology.gmail.com}
}
